php: short scalar cast

+++ web/documentserver-example/php/common.php
@@ -34,7 +34,7 @@ function guid() {
     if (function_exists('com_create_guid')) {
         return com_create_guid();
     }
-    mt_srand((double) microtime()*10000);  // optional for php 4.2.0 and up
+    mt_srand((float) microtime()*10000);  // optional for php 4.2.0 and up
     $charid = strtoupper(md5(uniqid(rand(), true)));
     $hyphen = chr(45);  // "-"
     $uuid = chr(123)  // "{"
